Draw unsigned shorts.
The GBA video memory can only be written 16bits at a time. This commit makes the main drawing routines draw pixels as (colour | (colour << 8)). I.e 2 pixels at a time.

+++ include/r_draw.h
@@ -74,8 +74,9 @@ typedef struct {
   const lighttable_t  *colormap;
 } draw_span_vars_t;
 
-typedef struct {
-  byte           *byte_topleft;
+typedef struct
+{
+  unsigned short *byte_topleft;
   int   byte_pitch;
 } draw_vars_t;
